+++ sso/verification/tests/test_models.py
@@ -10,7 +10,7 @@ from sso.verification.tests import factories
 
 @pytest.mark.django_db
 def test_verification_expired():
-    created = now() - timedelta(days=settings.VERIFICATION_EXPIRY_DAYS + 1)
+    created = now() - timedelta(minutes=settings.VERIFICATION_EXPIRY_MINUTES + 1)
     with freeze_time(created):
         verification_code = factories.VerificationCodeFactory()
 
@@ -19,7 +19,7 @@ def test_verification_expired():
 
 @pytest.mark.django_db
 def test_verification_not_expired():
-    created = now() - timedelta(days=settings.VERIFICATION_EXPIRY_DAYS - 1)
+    created = now() - timedelta(minutes=settings.VERIFICATION_EXPIRY_MINUTES - 1)
     with freeze_time(created):
         verification_code = factories.VerificationCodeFactory()